+++ packages/patternfly-4/react-core/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.7.0](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-core@3.6.0...@patternfly/react-core@3.7.0) (2019-04-22)
+
+
+### Features
+
+* **padding:** add pagesection no padding props ([#1818](https://github.com/patternfly/patternfly-react/issues/1818)) ([a472953](https://github.com/patternfly/patternfly-react/commit/a472953))
+
+
+
+
+
+# [3.6.0](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-core@3.5.0...@patternfly/react-core@3.6.0) (2019-04-22)
+
+
+### Features
+
+* **aboutmodal:** make productName optional ([#1797](https://github.com/patternfly/patternfly-react/issues/1797)) ([7384d26](https://github.com/patternfly/patternfly-react/commit/7384d26)), closes [#1714](https://github.com/patternfly/patternfly-react/issues/1714)
+
+
+
+
+
 # [3.5.0](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-core@3.4.1...@patternfly/react-core@3.5.0) (2019-04-22)